Relocate files due to new package structure
